QuestManager reorg - Move to subpackage
